Add upgrade check for NIC Switch agent
This commit adds an upgrade check for NIC switch agent
alerting operators to ensure relevant hosts have kernel >= 3.13.

This check is introduced due to recent cleanup in NIC switch
agent code.

As of U release, Supported CentOS major was updated to 8 so it
is not expected that any deployment will have hosts running
kernel < 3.13. This check is added as an extra precaution.

This check should be removed in 1-2 cycles. A TODO was added to
reflect that.

Change-Id: I34003b3c2f0ac23185d036c9e47dc1c8662d6ce7
Related-bug: #1841067