Start README.rst with a better title
Now that we are using gitea the contents of our README.rst are
more prominently displayed. Starting it with a "Team and repository
tags" title is a bit confusing. This change makes it start with the
name of the project instead.

While at it, also fix the intro and the CONTRIBUTING.rst mention,
for clarity.

Change-Id: Ibf4ae411d756744727a0aee24efdbdd561d0f522