Remove grenade-py3 job from Neutron CI
We have run neutron-grenade-multinode in both check and gate queues
already. This job is almost the same as grenade-py3. The only difference
from Neutron pov is that neutron-grenade-multinode is multinode job and
grenade-py3 is single node job.
So we agreed on last Neutron CI meeting [1], that there is no need to
run both of those jobs.

As grenade-py3 is part of integrated-gate-networking template together
with tempest-integrated-networking job this patch removes also
integrated-gate-networking template from neutron's zuul config and adds
only tempest-integrated-networking to both check and gate queues.

[1] http://eavesdrop.openstack.org/meetings/neutron_ci/2019/neutron_ci.2019-11-19-16.00.log.html#l-179

Change-Id: Ic2e7ca04929dc89a2e408465eb34ef3feec2e328