Add upgrade check for networking-ovn db version
When ovn code will be in neutron tree already, during the upgrade
we need to be sure that networking-ovn wasn't installed previously or
if it was installed, it's db version is the latest one.
Otherwise neutron's alembic migration script will not work properly as
it will check that ovn tables exists so it will skip creating them, even
if those tables aren't in proper shape (from latest networking-ovn
head).

Change-Id: I8c7b8df934c8b53ea5bc02dc8134db838f5e105a
Related-Blueprint: neutron-ovn-merge