Set ignore_basepython_conflict to true in tox.ini
Since dropping py2 support and adding a basepython setting to the global
tox testenv, tox complains about a basepython mismatch when the
installed python3 is python3.6 and the py37 environment is executed (or
vice versa). Setting ignore_basepython_conflict = True will avoid this
message and enforce the correct python version to be used.

Needs a bump in the minimum tox version supported.

See also [0].

[0] https://review.opendev.org/582392

Change-Id: I3d9beeaed104dcf12cfd97590b44c4003754b4e4