Install OVN for functional tests
This patch is changing the tools/configure_for_func_testing.sh script to
install OVN when setting up the environment for functional tests.

The "ovn" driver in master needs the latest released version of OVN
(2.12), unfortunately this version is not yet present in the ubuntu
repository so we need to compile it from source.

Partially-Implements: blueprint neutron-ovn-merge
Change-Id: I181da9981ec95da8cef36ba2bb667b5a4ec9db22
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>