Patch correct MAXFD for faster command execution
The backport of the patch patched the wrong subprocess MAXFD
parameter, as we need to patch the subprocess imported inside
greenlet. This has gone by unnoticed, as this function was seldomly
used because everything was going through rootwrap. With the new
agent setup used by us this is no longer the case and therefore
this function became the bottleneck.

With MAXFD being set on the correct subprocess command execution
is fast again.