don't clear skb mark when ovs is hw-offload enabled
skb mark is not supported when using ovs hw-offload and using it
breaks the vxlan offload.

This patch clear skb mark only if ovs hw-offload is disabled.
This should be fine as ovs with hw-offload runs on the
compute node (DVR is not supported), so clear the skb mark for
the qrouter is not needed.

Closes-Bug: #1855888

Change-Id: I71f45fcd9b7e7bdacaafc7fa96c775e88333ab48
(cherry picked from commit a75ec08ddb6a672ba685a11ba3f7df1569497723)