Pass correct port data to extension manager
The CRUD methods in the ML2 plugin must pass the data in the request
body to the extension manager, so it can be processed by all the ML2
extensions enabled in the deployment. The create_port_bulk method is
instead passing the dictionary generated after adding the port to the
DB. This will cause that fields in the request body intended for
extensions will not be processed.

# stein-only dependency for gate fix
Depends-On: https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/788965

Change-Id: I2fe8e2b98c05ba739e9e93b4bd5880eb97bdaef8
(cherry picked from commit 9c5156575dccf632195da2d25598bfce2823c3fa)
(cherry picked from commit efca646ba3afdd44497fc0e0235123a712b84b07)