Allow to kill keepalived state change monitor process
Usually Neutron stops neutron-keepalived-state-change-monitor process
gracefully with SIGTERM.
But in case if this will not stop process for some time, Neutron will
try to kill this process with SIGKILL (-9).
That was causing problem with rootwrap as kill filters for this process
allowed to send only "-15" to it.
Now it is possible to kill this process with "-9" too.

Conflicts:
    etc/neutron/rootwrap.d/l3.filters

Change-Id: Id019fa7649bd1158f9d56e63f8dad108d0ca8c1f
Closes-bug: #1860326
(cherry picked from commit d6fccd247f70abc84c8a480138e135717836c7b3)