[ovn] Agent liveness - allow time to propagate checks
Right now neutron-server bumps the nb_cfg parameter in NB_Global
table which needs to be propagated by northd to SB_Global,
processed by agents, and write it back into SB_Global.
This requires processing by neutron-server but unfortunatelly
the server checks straight away and many times the value read
is behind the expected value.

All this results in frequent false positives showing dead agents
when they are not.

This patch is relaxing the checks by allowing a difference of 1
between the read and expected values.

Change-Id: Id91481b690ad569c5dcfa5bd404f497f591d729d
Closes-Bug: 1860436
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>