[OVN] Add an interval between agents health checks
This patch adds a minimum interval between each agent health checks.

The way OVN checks for the agents liveness is by increasing a value in
the NB DB and waiting for it to be propagated to the SB DB but, this can
be costy if done many times too quickly. Therefore, a minimum interval
between each check is being added.

Closes-Bug: #1861092
Change-Id: If1f2d97e3a3a17f6744d546b3e8903bde55e83b9
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>