OVN Metadata agent gets OVSDB updates for only its Chassis
The metadata agent registers the Chassis table with ovsdb-server
and therefor gets database updates every time *any* Chassis is
updated--even if the update is just a liveness check that updates
nb_cfg.

This patch adds a condition so that metadata agent only gets updates
for the Chassis that it is running on.

Change-Id: I452b7de09312ecea621c4b448cc63f037cad9675
Related-bug: #1861092