[OVN] Reduce the number of watched tables in MetadataProxyHandler
This patch changes the SB IDL from the MetadataProxyHandler class to
only watch the two required tables for it to function.

The more tables it monitors, the bigger the burden is on the OVSDB to
keep all the in-memory replicas updated. In special, the Chassis table
can generate a lot of events due to the agents health check mechanism
in the OVN driver so, removing it will prevent many events from being
sent across the fabric.

Closes-Bug: #1862648
Change-Id: Ib4fb9b445ed7fe4bce7bc05ae1b8dca264324718
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>