[OVN] Devstack: Fix adding another compute node
In order to add another compute node to an existing DevStack setup
(multi-node) we need to set a few extra variables to the local.conf to
enable the OVN driver in neutron.

The patch is also adding a new neutron_plugin_create_nova_conf function
to the ovn_agent module (no-op) otherwise the DevStack setup will fail.

Change-Id: I4d53f7206f151dc7ffa51b4c8bd601279aa88a46
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>