[OVN] Use the Chassis_Private table for agents healthcheck
The core OVN team has introduced a new table called Chassis_Private to
avoid nb_cfg flooding when checking for the Chassis' status. The OVN
driver does rely on that mechanism for the agent liveness mechanism.

This patch makes use of this new table but it's also backward
compatible.

For more information, check the core OVN changes at:
https://patchwork.ozlabs.org/patch/1254394.

Closes-Bug: #1892477
Change-Id: Iea4263b852d1e3f81eb2557918ea3cbb7adb8016
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>