Set an absolute path in rootwrap filters_path
If a relative path is set in rootwrap filters_path, the rootwrap
command must be executed from the neutron root directory. Some
commands, like those execute by RootHelperProcess, do not have
a defined working directory. To avoid the problems present in the
related bug, an absolute path is set in the rootwrap configuration.

Related-Bug: #1862927

Change-Id: I02ce4d6b84242c42ec26954eeb9b776cb779bff4