Workaround in NetworkSegmentRange OVO until "project_id" migration
The "network_segment_range" service plugin API does not have
"tenant_id" as parameter. However, because the migration from
"tenant_id" to "project_id" did not finish, every OVO with
"project_id" will also have "tenant_id" as an extra field. This
extra field will be populated with the same value as the
"project_id" field.

If the Neutron API WSGI is used and "tenant_id" is not present in
the API (defined in neutron-lib), both "tenant_id" and "project_id"
parameters will be removed from the response.

This patch introduces a workaround for this OVO until the migration
is over.

More information about this patch can be found in the bug c#4.

Change-Id: I85a02d8b59e8a758826c110e00df84672fa93676
Closes-Bug: #1828205
(cherry picked from commit 3053caaa02adaadfffa53fb1070aaa60f49419e8)