Provide meaningful defaults for OVN/OVS_REPO_NAME
In local dev environments devstack may be configured to clone
a component from a git repo with a working directory, for example:
$HOME/src/ovn/.git

The original logic took the the base name without the .git extension
as the repo name to use - for this path that is the empty string,
which of course did not work.

This change supplies a meaningful default repo name for cases like this.

Change-Id: I3a2796f67d7f0634b1f25d44f4fa229d31ef9cc1