Ensure that fullstack job is run with python 2.7
On stable/queens branch we need to run fullstack job with
python 2.7. There wasn't basepython set explicity for
dsvm-fullstack and it seems that it was run on python 3.5 in
the gate.
So lets explicity set basepython for this job to avoid such problems.

This patch is needed for queens only as for rocky and newer versions
it was set explicity in [1].

[1] https://review.opendev.org/#/c/572973

Change-Id: Ic662f9b407c273beae676d0a8c93c48e624a2a96