Remove neutron-tempest-dvr job from CI
This is single node dvr job and we decided on the Neutron CI
meeting that we will remove it from both check and gate queues
and we will replace it with "neutron-tempest-dvr-ha-multinode-full"
which is similar but multinode job.

Job definition isn't removed as it is run in Tempest CI.

This patch also changes tox_envlist for the
"neutron-tempest-dvr-ha-multinode-full" job to run only networking
related tests intead of all tempest tests.

Change-Id: If78652478ec4720a4a0bc56e9de5a9186e33f57a