Change way of retrieving LSP in virtual_port functional tests
Sometimes we had a situation, where Logical_Switch_Port type
received by IDL had a different type, that type set in the
NBDB. We changed the way we retrieve the LS from DB, to not
loop over rows in table, but call db_find_rows() instead.
With this change after about 40 retries of functional tests we
don't spot the same issue again.

Change-Id: I07c081d1984b26a10a4d854d17117cfeaac7f8ac
Related-Bug: 1862618