Clean dvr fip gateway entry when fip agent gw port is deleted
In [1] there was introduced new db table which stored information about
which DVR L3 agent has got already floating ip gateway port. It was to
avoid race conditions and ensure that there is always only one such port
per network and per agent (host).
Unfortunately in [1] there was no added removal of correct record from
this db table so it was causing problems when such port had to be
recreated after it was already on the host and was deleted.

This patch adds removal of such entry from db when needed.

Closes-Bug: #1866336

[1] https://review.opendev.org/#/c/702547/

Change-Id: I56efd1b9f09c0449ce531a185fcf4db353f99fe1