Mark security group fullstack tests as stable
It has been some time since we marked the security group tests in
fullstack as unstable.  Now that mark is impeding the fullstack tests
for stateless security groups.

Based on testing it appears that the unstable tests are specific to the
linuxbridge driver, which is not really very much supported these days.
Testing for that driver is dropped, since it is better to at least test
for the other scenarios than nothing at all.

Change-Id: Ib1ef0763496aa536ba80f1c34c64d8a6c313fd0b
Needed-By: https://review.opendev.org/710252
Related-bug: #1779328