Fix _notify_l3_agent_port_update when there is no binding host
When _notify_l3_agent_port_update is called, there may be the case that
portbindings.HOST_ID key is not in new_port or original_port dict. It is
like that e.g. in some unit tests of api extensions. So lets use get()
method to get it. The case when host_id is None is already handled
properly in this method so that is fine to do.

Change-Id: I490506cc6c9b18a3ca12e474f9ee9f8f3712ce7b