Add fip_pf_description extension to be supported by service plugin
In [1] new api extension "fip_pf_description" was introduced but it
wasn't added to the list of supported extensions by port_forwarding
service plugin.
Because of that "description" attribute was unknown for the
port_forwarding resource.
Now this new api extension is added and supported by pf plugin.

[1] https://review.opendev.org/#/c/670930/

Closes-Bug: #1866560
Partially-implements: bp/fip-pf-description

Depends-On: https://review.opendev.org/#/c/711856/
Change-Id: Ibf42a4d276d0141d66ae6e88aa9fbc291eaa4f82