[OVN] Switch release jobs to OVN 20.03 (v20.03.0 tag)
OVN 20.03 (v20.03.0 tag) is the new stable released version of OVN which
is compatible with the OVS 2.13 version, also bumped in this review.

The patch also updates the use_new_ovn_repository to strip all
non-numeric characters of the branch name when comparing the versions,
that way we can compare branch-<version> against v<version> version
formats.

Change-Id: I06f06fbc42c625ad2deab03a92238ae5e68d9033
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>